// src/types/api.ts
export interface components {
  schemas: {
// ...
    anime: {
      /** @description MyAnimeList ID */
      mal_id?: number;
      /** @description MyAnimeList URL */
      url?: string;
      images?: components['schemas']['anime_images'];
      trailer?: components['schemas']['trailer_base'];
      /** @description Whether the entry is pending approval on MAL or not */
      approved?: boolean;
      /** @description All titles */
      titles?: components['schemas']['title'][];
      /**
       * @deprecated
       * @description Title
       */
      title?: string;
      /**
       * @deprecated
       * @description English Title
       */
      title_english?: string | null;
      /**
       * @deprecated
       * @description Japanese Title
       */
      title_japanese?: string | null;
      /**
       * @deprecated
       * @description Other Titles
       */
      title_synonyms?: string[];
      /**
       * @description Anime Type
       * @enum {string|null}
       */
      type?: 'TV' | 'OVA' | 'Movie' | 'Special' | 'ONA' | 'Music' | null;
      /** @description Original Material/Source adapted from */
      source?: string | null;
      /** @description Episode count */
      episodes?: number | null;
      /**
       * @description Airing status
       * @enum {string|null}
       */
      status?: 'Finished Airing' | 'Currently Airing' | 'Not yet aired' | null;
      /** @description Airing boolean */
      airing?: boolean;
      aired?: components['schemas']['daterange'];
      /** @description Parsed raw duration */
      duration?: string | null;
      /**
       * @description Anime audience rating
       * @enum {string|null}
       */
      rating?:
        | 'G - All Ages'
        | 'PG - Children'
        | 'PG-13 - Teens 13 or older'
        | 'R - 17+ (violence & profanity)'
        | 'R+ - Mild Nudity'
        | 'Rx - Hentai'
        | null;
      /**
       * Format: float
       * @description Score
       */
      score?: number | null;
      /** @description Number of users */
      scored_by?: number | null;
      /** @description Ranking */
      rank?: number | null;
      /** @description Popularity */
      popularity?: number | null;
      /** @description Number of users who have added this entry to their list */
      members?: number | null;
      /** @description Number of users who have favorited this entry */
      favorites?: number | null;
      /** @description Synopsis */
      synopsis?: string | null;
      /** @description Background */
      background?: string | null;
      /**
       * @description Season
       * @enum {string|null}
       */
      season?: 'summer' | 'winter' | 'spring' | 'fall' | null;
      /** @description Year */
      year?: number | null;
      broadcast?: components['schemas']['broadcast'];
      producers?: components['schemas']['mal_url'][];
      licensors?: components['schemas']['mal_url'][];
      studios?: components['schemas']['mal_url'][];
      genres?: components['schemas']['mal_url'][];
      explicit_genres?: components['schemas']['mal_url'][];
      themes?: components['schemas']['mal_url'][];
      demographics?: components['schemas']['mal_url'][];
    };
// ...
  };
// ...
}